# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API circleoflinks",
        "description": "API de red social de profecionales ",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/circleoflinks-app/public"
        }
    ],
    "paths": {
...
        "/api/v1/userofertalaboral": {
            "get": {
                "tags": [
                    "User oferta laboral"
                ],
                "summary": "Se muestran los registros de user oferta laborales del user ",
                "description": "show statususer",
                "operationId": "30ab2f021ddc4b375829060d07e1d4fc",
                "responses": {
                    "200": {
                        "description": "descripción o el nombre del código de la petición",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "user_id": {
                                                        "type": "string",
                                                        "example": "description"
                                                    },
                                                    "oferta_laboral_id": {
                                                        "type": "string",
                                                        "example": "description"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "User oferta laboral"
                ],
                "summary": "Se actualiza el registro que es especificado en el campo id ",
                "description": "update user Oferta Laboral",
                "operationId": "62ff96a1493f76a387c12dd2c6cd2368",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "oferta_laboral_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "idescription",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descripción o el nombre del código de la petición",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Aderson Felix"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "angelshamael@gmail.com"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "descripción o el nombre del código de la petición",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "angelshamael@gmail.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/userofertalaboral/{id}": {
            "get": {
                "tags": [
                    "User oferta laboral"
                ],
                "summary": "Se muestra el registro de las ofertas laborales del user ",
                "description": "show statususer",
                "operationId": "2b991733cc08505865f2572483472dc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descripción o el nombre del código de la petición",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "user_id": {
                                                        "type": "string",
                                                        "example": "description"
                                                    },
                                                    "oferta_laboral_id": {
                                                        "type": "string",
                                                        "example": "description"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}